requirements.txt
pdf2image
streamlit
google-generativeai
python-dotenv
pdf2image
Pillow
requests
pymupdf